Revert "Skill Panel - Added 'lit up green' exit flag icon for when save requirement is met" (commit 397dc6c).
We cannot add icons yet as this will upset UI for users of both 12.14 and CE
(some users wish to use both from the same directory - we should
continue to support this as long as possible).

In due course, we may be able to add new icons by supplementing with an
additional sheet (with graceful handling if it's missing), but for now
let's just work with what we already have.